emotion module reformed, results are returned with EmotionResult objects
