--- projectoxford/emotion/__init__.py
@@ -1,16 +1,0 @@
-try:
-    import numpy
-except ImportError:
-    raise ImportError('Package numpy is not installed.')
-
-
-try:
-    import matplotlib.pyplot
-except ImportError:
-    raise ImportError('Package matplotlib is not installed.')
-
-
-try:
-    import cv2
-except ImportError:
-    raise ImportError('Package opencv for python is not installed')
