+++ projectoxford/endpoints.py
@@ -1,2 +1,1 @@
 EMOTION_ENDPOINT = 'https://api.projectoxford.ai/emotion/v1.0/recognize'
-MAX_NUM_RETRIES = 10
